+++ src/content/coupang/categoryHandler/categoryHandlerCosmetic.tsx
