+++ src/content/index.tsx
@@ -6,9 +6,9 @@ import { handleStyleMessage } from "./messageHandlers/styleMessageHandler";
 import { handleCursorMessage } from "./messageHandlers/cursorMessageHandler";
 import { handleModalMessage } from "./messageHandlers/modalMessageHandler";
 import { processImages } from "./imageHandlers/imageProcessor";
-import { MountCartSummaryApp } from "./coupang/cartSummary";
-import { checkCategoryCosmeticAndRender } from "./coupang/categoryHandlerCosmetic";
-import { checkCategoryFoodAndRender } from "./coupang/categoryHandlerFood";
+import { MountCartSummaryApp } from "./coupang/CartSummary";
+import { checkCategoryCosmeticAndRender } from "./coupang/categoryHandler/categoryHandlerCosmetic";
+import { checkCategoryFoodAndRender } from "./coupang/categoryHandler/categoryHandlerFood";
 import { renderCouponComponent } from "./coupang/renderCouponComponent";
 import { renderInfoComponent } from "../content/coupang/renderInfoComponent";
 import { initDomObserver } from "./observers/domObserver";
